FWMT-405 Code review changes
