+++ src/main/java/uk/gov/ons/fwmt/job_service/service/impl/JobServiceImpl.java
@@ -1,9 +1,15 @@
 package uk.gov.ons.fwmt.job_service.service.impl;
 
-import lombok.extern.slf4j.Slf4j;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+
+import lombok.extern.slf4j.Slf4j;
 import uk.gov.ons.fwmt.job_service.data.csv_parser.CSVParseResult;
 import uk.gov.ons.fwmt.job_service.data.csv_parser.UnprocessedCSVRow;
 import uk.gov.ons.fwmt.job_service.data.dto.SampleSummaryDTO;
@@ -12,61 +18,44 @@ import uk.gov.ons.fwmt.job_service.data.legacy_ingest.LegacySampleIngest;
 import uk.gov.ons.fwmt.job_service.exceptions.ExceptionCode;
 import uk.gov.ons.fwmt.job_service.exceptions.types.InvalidFileNameException;
 import uk.gov.ons.fwmt.job_service.exceptions.types.MediaTypeNotSupportedException;
-import uk.gov.ons.fwmt.job_service.rest.FieldPeriodResourceService;
-import uk.gov.ons.fwmt.job_service.rest.JobResourceService;
-import uk.gov.ons.fwmt.job_service.rest.UserResourceService;
 import uk.gov.ons.fwmt.job_service.service.CSVParsingService;
 import uk.gov.ons.fwmt.job_service.service.FileIngestService;
 import uk.gov.ons.fwmt.job_service.service.JobService;
-import uk.gov.ons.fwmt.job_service.service.totalmobile.TMJobConverterService;
-import uk.gov.ons.fwmt.job_service.service.totalmobile.TMService;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
 
 @Slf4j
 @Service
 public class JobServiceImpl implements JobService {
   private FileIngestService fileIngestService;
   private CSVParsingService csvParsingService;
-  private TMJobConverterService tmJobConverterService;
-  private TMService tmService;
-  private UserResourceService userResourceService;
-  private JobResourceService jobResourceService;
-  private FieldPeriodResourceService fieldPeriodResourceService;
-  private JobProcessServiceImpl jobProcessService;
+  private JobProcessorServiceImpl jobProcessService;
 
   @Autowired
   public JobServiceImpl(
       FileIngestService fileIngestService,
       CSVParsingService csvParsingService,
-      TMJobConverterService tmJobConverterService,
-      TMService tmService,
-      UserResourceService userResourceService,
-      JobResourceService jobResourceService,
-      FieldPeriodResourceService fieldPeriodResourceService,
-      JobProcessServiceImpl jobProcessService) {
+      JobProcessorServiceImpl jobProcessService) {
     this.fileIngestService = fileIngestService;
     this.csvParsingService = csvParsingService;
-    this.tmJobConverterService = tmJobConverterService;
-    this.tmService = tmService;
-    this.userResourceService = userResourceService;
-    this.jobResourceService = jobResourceService;
-    this.fieldPeriodResourceService = fieldPeriodResourceService;
     this.jobProcessService = jobProcessService;
   }
 
   @Override
-  public SampleSummaryDTO validateSampleFile(MultipartFile file)
+  public SampleSummaryDTO processSampleFile(MultipartFile file)
           throws IOException, InvalidFileNameException, MediaTypeNotSupportedException{
+    SampleSummaryDTO sampleSummaryDTO = validateSampleFile(file);
+
+    // This is an async call
+    jobProcessService.processSampleFile(file);
+
+    return sampleSummaryDTO;
+  }
+  
+  private SampleSummaryDTO validateSampleFile(MultipartFile file) throws InvalidFileNameException, MediaTypeNotSupportedException, IOException{
     FileIngest fileIngest = fileIngestService.ingestSampleFile(file);
     Iterator<CSVParseResult<LegacySampleIngest>> csvRowIterator = csvParsingService.parseLegacySample(fileIngest.getReader(), fileIngest.getFilename().getTla());
 
     int parsed = 0;
     List<UnprocessedCSVRow> unprocessed = new ArrayList<>();
-
 
     while (csvRowIterator.hasNext()) {
       CSVParseResult<LegacySampleIngest> row = csvRowIterator.next();
@@ -78,12 +67,6 @@ public class JobServiceImpl implements JobService {
       parsed++;
     }
 
-    jobProcessService.processSampleFile(file);
-
-    System.out.println("sending message");
-
-    // construct reply
     return new SampleSummaryDTO(file.getOriginalFilename(), parsed, unprocessed);
   }
-
 }
